requirements.txt
langchain
langchain-community
langchain_openai
openai
python-dotenv
chromadb